# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/clawbacks:
    get:
      description: Requires clawback:view. Scoped; filters status/rep_id/sale_id.
      operationId: ClawbackController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - pending
              - applied
        - name: rep_id
          required: false
          in: query
          schema:
            type: string
        - name: sale_id
          required: false
          in: query
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: &a25
        - bearer: []
      summary: List clawbacks
      tags: &a26
        - Clawback
    post:
      description: Requires clawback:create. Targets a PAID sale_item; amount defaults
        to the exact frozen amount (rate + incentive). Flat, per-item, no date
        math; the snapshot is never edited.
      operationId: ClawbackController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClawbackDto"
      responses:
        "201":
          description: ""
      security: *a25
      summary: Enter a clawback
      tags: *a26
  /v1/clawbacks/{id}:
    get:
      description: Requires clawback:view (scoped).
      operationId: ClawbackController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a25
      summary: Get a clawback
      tags: *a26
  /v1/expense-reports:
    post:
      description: Requires expenses:create. Any user may submit (own by default). km
        items compute their amount; non-km items require a receipt per the
        category config. The pay period is derived from week_start.
      operationId: ExpenseReportsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportDto"
      responses:
        "201":
          description: ""
      security: &a27
        - bearer: []
      summary: Submit a weekly expense report
      tags: &a28
        - Expenses
    get:
      description: Requires expenses:view. Scoped (own/roster/all); filters
        status/rep/client/period/date.
      operationId: ExpenseReportsController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - draft
              - submitted
              - approved
              - rejected
              - sent_back
        - name: rep_id
          required: false
          in: query
          description: Filter by rep.
          schema:
            type: string
        - name: client_id
          required: false
          in: query
          description: Filter by a client tagged on any item.
          schema:
            type: string
        - name: pay_period_id
          required: false
          in: query
          description: Filter by pay period.
          schema:
            type: string
        - name: from
          required: false
          in: query
          description: Reports whose week_start ≥ this date.
          schema:
            example: 2026-03-01
            type: string
        - name: to
          required: false
          in: query
          description: Reports whose week_start ≤ this date.
          schema:
            example: 2026-03-31
            type: string
      responses:
        "200":
          description: ""
      security: *a27
      summary: List expense reports
      tags: *a28
  /v1/expense-reports/{id}:
    get:
      description: Requires expenses:view (scoped).
      operationId: ExpenseReportsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a27
      summary: Get an expense report
      tags: *a28
    patch:
      description: Requires expenses:edit. Editable pre-approval; once approved, only
        a Super Admin may edit. Supplying items replaces the report lines
        wholesale.
      operationId: ExpenseReportsController_edit_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReportDto"
      responses:
        "200":
          description: ""
      security: *a27
      summary: Edit an expense report
      tags: *a28
  /v1/expense-reports/{id}/approve:
    post:
      description: Requires expenses:approve. decision = approve | reject | send_back.
      operationId: ExpenseReportsController_review_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDto"
      responses:
        "201":
          description: ""
      security: *a27
      summary: Review an expense report
      tags: *a28
  /v1/expense-field-configs:
    get:
      description: Requires expenses:view.
      operationId: ExpenseFieldConfigsController_list_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: &a29
        - bearer: []
      summary: List expense category configs
      tags: &a30
        - Expenses
    post:
      description: Requires expenses:edit. Sets label, requires_receipt, is_active for
        a category key.
      operationId: ExpenseFieldConfigsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFieldConfigDto"
      responses:
        "201":
          description: ""
      security: *a29
      summary: Add / configure an expense category
      tags: *a30
  /v1/expense-exports:
    get:
      description: Requires expenses:view.
      operationId: ExpenseExportsController_list_v1
      parameters: []
      responses:
        "200":
          description: ""
      security: &a31
        - bearer: []
      summary: List expense exports
      tags: &a32
        - Expenses
    post:
      description: Requires expenses:export. Records the request with a stubbed
        file_url (generation deferred).
      operationId: ExpenseExportsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExportDto"
      responses:
        "201":
          description: ""
      security: *a31
      summary: Generate an expense export
      tags: *a32
...
  /v1/documents:
    post:
      description: Requires documents:create. Stores a stubbed original_file_url;
        status draft, owner = caller.
      operationId: DocumentsController_upload_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentDto"
      responses:
        "201":
          description: ""
      security: &a41
        - bearer: []
      summary: Upload a document
      tags: &a42
        - Documents & E-Signature
    get:
      description: Requires documents:view. Scoped to owned or shared-with
        (Admin/Super Admin see all).
      operationId: DocumentsController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - draft
              - shared
              - partially_signed
              - completed
              - declined
        - name: doc_type
          required: false
          in: query
          schema:
            type: string
            enum:
              - compensation_agreement
              - rate_notice
              - equipment
              - other
      responses:
        "200":
          description: ""
      security: *a41
      summary: List documents
      tags: *a42
  /v1/documents/{id}:
    get:
      description: Requires documents:view. 404 if not visible to the caller.
      operationId: DocumentsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a41
      summary: Get a document (requests + per-signer status)
      tags: *a42
  /v1/documents/{id}/signature-requests:
    post:
      description: Requires documents:create AND ownership of the document (else 403).
        Recipients become the shared-with set.
      operationId: DocumentsController_requestSignature_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignatureRequestDto"
      responses:
        "201":
          description: ""
      security: *a41
      summary: Share + request a signature from one or many recipients
      tags: *a42
  /v1/signature-requests/{id}/sign:
    post:
      description: Authenticated; the caller must be a recipient of this request (else
        403). decision = sign | decline. Records the signature event (method,
        IP, timestamp) and stores a stubbed signed copy.
      operationId: SignatureRequestsController_sign_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignDto"
      responses:
        "201":
          description: ""
      security: &a43
        - bearer: []
      summary: Sign or decline a signature request
      tags: &a44
        - Documents & E-Signature
  /v1/signature-requests/{id}/cancel:
    post:
      description: Authenticated; only the requester, the document owner, or an admin
        may cancel (else 403).
      operationId: SignatureRequestsController_cancel_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      security: *a43
      summary: Cancel a signature request
      tags: *a44
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateClawbackDto:
      type: object
      properties:
        sale_item_id:
          type: string
          description: The cancelled, already-paid sale_item to claw back.
        reason:
          type: string
          example: Customer cancelled within contract window.
        reported_date:
          type: string
          example: 2026-03-15
          description: Client-reported cancellation date (record only; drives no logic).
        amount:
          type: string
          example: "30.00"
          description: Override the snapshot-derived amount (editable with permission).
            Decimal string, never float.
      required:
        - sale_item_id
        - reason
        - reported_date
    KmStopInput:
      type: object
      properties:
        stop_order:
          type: number
          example: 0
          description: Order of this stop along the route (0-based).
        address:
          type: string
          example: 123 Main St, Winnipeg
        lat:
          type: string
          example: "49.895100"
          description: Latitude (decimal string).
        lng:
          type: string
          example: "-97.138400"
          description: Longitude (decimal string).
      required:
        - stop_order
        - address
        - lat
        - lng
    KmLogInput:
      type: object
      properties:
        trip_type:
          type: string
          enum:
            - single
            - round
          example: round
        total_km:
          type: string
          example: "130.00"
          description: Route's total driven distance in km (decimal string).
        stops:
          type: array
          items:
            $ref: "#/components/schemas/KmStopInput"
      required:
        - trip_type
        - total_km
        - stops
    ExpenseItemInput:
      type: object
      properties:
        category:
          type: string
          enum:
            - km
            - meals
            - hotel
            - flight
            - rental
            - gas
            - other
          example: meals
        client_id:
          type: string
          description: Optional client this expense is attributed to.
        expense_date:
          type: string
          example: 2026-03-10
        amount:
          type: string
          example: "42.50"
          description: Amount (decimal string). Required for non-km items; ignored for km
            (computed).
        description:
          type: string
          example: Lunch with client
        receipt_url:
          type: string
          description: Object-storage reference; mandatory per the category config.
        km:
          description: Required for `km` items; forbidden otherwise.
          allOf:
            - $ref: "#/components/schemas/KmLogInput"
      required:
        - category
        - expense_date
        - description
    CreateReportDto:
      type: object
      properties:
        week_start:
          type: string
          example: 2026-03-09
          description: Week start (governs the pay period).
        week_end:
          type: string
          example: 2026-03-15
        rep_id:
          type: string
          description: Rep this report is for; defaults to the submitter’s linked rep.
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseItemInput"
      required:
        - week_start
        - week_end
        - items
    UpdateReportDto:
      type: object
      properties:
        week_start:
          type: string
          example: 2026-03-09
        week_end:
          type: string
          example: 2026-03-15
        items:
          description: Replaces all lines when provided.
          type: array
          items:
            $ref: "#/components/schemas/ExpenseItemInput"
    ReviewDto:
      type: object
      properties:
        decision:
          type: string
          enum:
            - approve
            - reject
            - send_back
        note:
          type: string
          description: Optional note (reason for rejection / send-back).
      required:
        - decision
    CreateFieldConfigDto:
      type: object
      properties:
        category_key:
          type: string
          example: parking
          description: Snake_case category key (unique).
        label:
          type: string
          example: Parking
        requires_receipt:
          type: boolean
          example: true
          description: Whether items in this category require a receipt.
        is_active:
          type: boolean
          example: true
          default: true
      required:
        - category_key
        - label
        - requires_receipt
    CreateExportDto:
      type: object
      properties:
        format:
          type: string
          enum:
            - pdf
            - excel
          example: excel
        client_id:
          type: string
          description: Restrict to one client.
        pay_period_id:
          type: string
          description: Restrict to one pay period.
      required:
        - format
...
    CreateDocumentDto:
      type: object
      properties:
        title:
          type: string
          example: Compensation Agreement 2026
        doc_type:
          type: string
          enum:
            - compensation_agreement
            - rate_notice
            - equipment
            - other
          example: compensation_agreement
      required:
        - title
        - doc_type
    CreateSignatureRequestDto:
      type: object
      properties:
        recipient_user_ids:
          description: User ids to request a signature from (≥1).
          type: array
          items:
            type: string
        message:
          type: string
          example: Please review and sign by Friday.
        due_date:
          type: string
          example: 2026-06-30
          description: Optional due date (record only).
      required:
        - recipient_user_ids
    SignDto:
      type: object
      properties:
        decision:
          type: string
          enum:
            - sign
            - decline
        method:
          type: string
          example: typed
          description: Signature method (typed / drawn / …).
      required:
        - decision
...